+++ aquadoggo/src/schema/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod schema_provider;
+
+pub use schema_provider::SchemaProvider;